requirements.txt
python-dotenv==0.21.1
selenium==4.8.0
webdriver_manager==3.8.5
